Make AutoLinkPlugin matchers export
