+++ packages/lexical-react/src/LexicalAutoLinkPlugin.ts
@@ -39,6 +39,38 @@ type LinkMatcherResult = {
 
 export type LinkMatcher = (text: string) => LinkMatcherResult | null;
 
+const URL_MATCHER =
+  /((https?:\/\/(www\.)?)|(www\.))[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)/;
+
+export function urlLinkMatcher(text: string) {
+  const match = URL_MATCHER.exec(text);
+  if (match === null) {
+    return null;
+  }
+  const fullMatch = match[0];
+  return {
+    index: match.index,
+    length: fullMatch.length,
+    text: fullMatch,
+    url: fullMatch.startsWith('http') ? fullMatch : `https://${fullMatch}`,
+  };
+}
+
+const EMAIL_MATCHER =
+  /(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))/;
+
+export function emailLinkMatcher(text: string) {
+  const match = EMAIL_MATCHER.exec(text);
+  return (
+    match && {
+      index: match.index,
+      length: match[0].length,
+      text: match[0],
+      url: `mailto:${match[0]}`,
+    }
+  );
+}
+
 function findFirstMatch(
   text: string,
   matchers: Array<LinkMatcher>,
